+++ src/main/java/subway/line/application/dto/SectionDeleteRequest.java
@@ -1,0 +1,11 @@
+package subway.line.application.dto;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class SectionDeleteRequest {
+    private Long stationId;
+    private Long lineId;
+}
