--- src/main/java/subway/line/controller/LineController.java
@@ -1,78 +1,0 @@
-package subway.line.controller;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
-import subway.line.dto.LineCreateRequest;
-import subway.line.dto.LineModifyRequest;
-import subway.line.dto.LineResponse;
-import subway.line.dto.SectionCreateRequest;
-import subway.line.dto.SectionDeleteRequest;
-import subway.line.service.LineService;
-
-import java.net.URI;
-import java.util.List;
-
-@RestController
-@RequestMapping("/lines")
-@RequiredArgsConstructor
-public class LineController {
-
-    private final LineService lineService;
-
-    @PostMapping
-    public ResponseEntity<LineResponse> createLine(@RequestBody LineCreateRequest lineRequest) {
-        LineResponse line = lineService.createLine(lineRequest);
-        return ResponseEntity.created(URI.create("/lines/" + line.getId())).body(line);
-    }
-
-    @GetMapping
-    public ResponseEntity<List<LineResponse>> retrieveLines() {
-        List<LineResponse> lineResponses = lineService.findAll();
-        return ResponseEntity.ok().body(lineResponses);
-    }
-
-    @GetMapping("/{id}")
-    public ResponseEntity<LineResponse> retrieveLine(@PathVariable Long id) {
-        LineResponse lineResponse = lineService.findById(id);
-        return ResponseEntity.ok().body(lineResponse);
-    }
-
-    @PutMapping("/{id}")
-    public ResponseEntity<Void> modifyLine(@PathVariable Long id, @RequestBody LineModifyRequest request) {
-        lineService.updateLine(id, request);
-        return ResponseEntity.ok().build();
-    }
-
-    @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteLine(@PathVariable Long id) {
-        lineService.deleteById(id);
-        return ResponseEntity.noContent().build();
-    }
-
-    @PostMapping("/{id}/sections")
-    public ResponseEntity<Void> appendSection(@PathVariable(name = "id") Long lineId, @RequestBody SectionCreateRequest request) {
-        lineService.appendSection(lineId, request);
-        return ResponseEntity.ok().build();
-    }
-
-    @DeleteMapping("/{id}/sections")
-    public ResponseEntity<Void> deleteSection(@PathVariable(name = "id") Long lineId, @RequestParam(name = "stationId") Long stationId) {
-        SectionDeleteRequest request = SectionDeleteRequest.builder()
-                .stationId(stationId)
-                .lineId(lineId)
-                .build();
-        lineService.deleteSection(request);
-        return ResponseEntity.noContent().build();
-    }
-
-
-}
