+++ src/main/java/subway/line/domain/LineRepository.java
@@ -1,0 +1,8 @@
+package subway.line.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import subway.line.domain.Line;
+
+public interface LineRepository extends JpaRepository<Line, Long> {
+
+}
