--- src/main/java/subway/line/dto/SectionDeleteRequest.java
@@ -1,11 +1,0 @@
-package subway.line.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class SectionDeleteRequest {
-    private Long stationId;
-    private Long lineId;
-}
