+++ src/main/java/subway/station/application/dto/StationRequest.java
@@ -1,0 +1,13 @@
+package subway.station.application.dto;
+
+import lombok.Getter;
+
+import javax.validation.constraints.NotBlank;
+
+@Getter
+public class StationRequest {
+
+    private final String NAME_NOT_BLANK_MESSAGE = "역 생성을 위해 역 이름은 필수 값 입니다.";
+    @NotBlank(message = NAME_NOT_BLANK_MESSAGE)
+    private String name;
+}
