+++ src/main/java/subway/station/application/dto/StationResponse.java
@@ -1,0 +1,40 @@
+package subway.station.application.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
+import subway.station.domain.Station;
+
+import java.util.List;
+import java.util.stream.Collectors;
+
+@ToString
+@Getter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class StationResponse {
+
+    private Long id;
+
+    private String name;
+
+    public static StationResponse from(Station station) {
+        return StationResponse.builder()
+                .id(station.getId())
+                .name(station.getName())
+                .build();
+    }
+
+    public static List<StationResponse> from(List<Station> stations) {
+        return stations.stream()
+                .map(station -> StationResponse.builder()
+                        .id(station.getId())
+                        .name(station.getName())
+                        .build())
+                .collect(Collectors.toList());
+
+    }
+}
