--- src/main/java/subway/station/controller/StationController.java
@@ -1,41 +1,0 @@
-package subway.station.controller;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-import subway.station.dto.StationRequest;
-import subway.station.dto.StationResponse;
-import subway.station.service.StationService;
-
-import java.net.URI;
-import java.util.List;
-
-@RestController
-@RequestMapping("/stations")
-@RequiredArgsConstructor
-public class StationController {
-    private final StationService stationService;
-
-    @PostMapping
-    public ResponseEntity<StationResponse> createStation(@RequestBody StationRequest stationRequest) {
-        StationResponse station = stationService.saveStation(stationRequest);
-        return ResponseEntity.created(URI.create("/stations/" + station.getId())).body(station);
-    }
-
-    @GetMapping
-    public ResponseEntity<List<StationResponse>> showStations() {
-        return ResponseEntity.ok().body(stationService.findAll());
-    }
-
-    @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteStation(@PathVariable Long id) {
-        stationService.deleteById(id);
-        return ResponseEntity.noContent().build();
-    }
-}
