+++ src/main/java/subway/station/domain/StationRepository.java
@@ -1,0 +1,6 @@
+package subway.station.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface StationRepository extends JpaRepository<Station, Long> {
+}
