+++ src/test/java/subway/acceptance/line/SectionFixture.java
@@ -1,8 +1,8 @@
 package subway.acceptance.line;
 
-import subway.line.model.Line;
-import subway.line.model.Section;
-import subway.station.model.Station;
+import subway.line.domain.Line;
+import subway.line.domain.Section;
+import subway.station.domain.Station;
 
 import java.util.HashMap;
 import java.util.Map;
