+++ src/test/java/subway/unit/line/LineServiceMockTest.java
@@ -6,14 +6,14 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import subway.line.dto.LineModifyRequest;
-import subway.line.dto.SectionCreateRequest;
-import subway.line.dto.SectionDeleteRequest;
-import subway.line.model.Line;
-import subway.line.repository.LineRepository;
-import subway.line.service.LineService;
-import subway.station.model.Station;
-import subway.station.service.StationService;
+import subway.line.application.LineService;
+import subway.line.application.dto.LineModifyRequest;
+import subway.line.application.dto.SectionCreateRequest;
+import subway.line.application.dto.SectionDeleteRequest;
+import subway.line.domain.Line;
+import subway.line.domain.LineRepository;
+import subway.station.application.StationService;
+import subway.station.domain.Station;
 
 import javax.persistence.EntityNotFoundException;
 import java.util.Optional;
