+++ src/test/java/subway/unit/line/LineServiceTest.java
@@ -6,14 +6,14 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.transaction.annotation.Transactional;
 import subway.exception.SubwayNotFoundException;
-import subway.line.dto.LineModifyRequest;
-import subway.line.dto.SectionCreateRequest;
-import subway.line.dto.SectionDeleteRequest;
-import subway.line.model.Line;
-import subway.line.repository.LineRepository;
-import subway.line.service.LineService;
-import subway.station.model.Station;
-import subway.station.repository.StationRepository;
+import subway.line.application.dto.LineModifyRequest;
+import subway.line.application.dto.SectionCreateRequest;
+import subway.line.application.dto.SectionDeleteRequest;
+import subway.line.domain.Line;
+import subway.line.domain.LineRepository;
+import subway.line.application.LineService;
+import subway.station.domain.Station;
+import subway.station.domain.StationRepository;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
