+++ src/test/java/subway/unit/line/LineTest.java
@@ -4,9 +4,9 @@ package subway.unit.line;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import subway.acceptance.line.SectionFixture;
-import subway.line.model.Line;
-import subway.line.model.Section;
-import subway.station.model.Station;
+import subway.line.domain.Line;
+import subway.line.domain.Section;
+import subway.station.domain.Station;
 
 import java.util.List;
