+++ src/test/java/subway/unit/path/PathServiceMockTest.java
@@ -6,15 +6,15 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import subway.line.model.Line;
-import subway.line.model.Section;
-import subway.line.service.LineService;
-import subway.path.component.PathFinder;
-import subway.path.dto.PathRetrieveResponse;
-import subway.path.service.PathService;
-import subway.station.dto.StationResponse;
-import subway.station.model.Station;
-import subway.station.service.StationService;
+import subway.line.domain.Line;
+import subway.line.domain.Section;
+import subway.line.application.LineService;
+import subway.path.application.PathFinder;
+import subway.path.application.dto.PathRetrieveResponse;
+import subway.path.application.PathService;
+import subway.station.application.dto.StationResponse;
+import subway.station.domain.Station;
+import subway.station.application.StationService;
 
 import java.util.List;
